chore: all components organised and in directories
