+++ src/Components/Categories/Categories.js
@@ -3,7 +3,7 @@ import { Droppable, Draggable } from 'react-beautiful-dnd'
 import { useAppState, useAppDispatch } from '../../state-context'
 
 import { BlockWrapper } from '../styles/ContentBlocks'
-import { AutoFillCategories } from '../AutoFill'
+import { AutoFillCategories } from '../Autofill'
 
 const Categories = props => {
   const { categories } = useAppState()
