+++ src/Components/Preview/CodePreview.js
@@ -1,9 +1,9 @@
 import React from 'react'
-import { useAppState } from '../state-context'
+import { useAppState } from '../../state-context'
 import styled from 'styled-components'
 
 import { PrismLight as SyntaxHighlighter } from 'react-syntax-highlighter'
-import prism from '../lib/Prism/theme'
+import prism from '../../lib/Prism/theme'
 import html from 'react-syntax-highlighter/dist/esm/languages/prism/markup'
 
 const CodeEditor = styled.div`
